+++ tests/http/test_stream.c
@@ -68,7 +68,7 @@ AVS_UNIT_TEST(http, init) {
     AVS_UNIT_ASSERT_EQUAL(stream->status, 0);
     AVS_UNIT_ASSERT_EQUAL(stream->redirect_count, 0);
     AVS_UNIT_ASSERT_NULL(stream->user_headers);
-    avs_unit_mocksock_expect_shutdown(socket);
+    avs_unit_mocksock_expect_mid_close(socket);
     AVS_UNIT_ASSERT_SUCCESS(avs_stream_cleanup((avs_stream_t **) &stream));
     avs_http_free(client);
 }
@@ -126,7 +126,7 @@ AVS_UNIT_TEST(http, init_https_auth1) {
     AVS_UNIT_ASSERT_EQUAL(stream->status, 0);
     AVS_UNIT_ASSERT_EQUAL(stream->redirect_count, 0);
     AVS_UNIT_ASSERT_NULL(stream->user_headers);
-    avs_unit_mocksock_expect_shutdown(socket);
+    avs_unit_mocksock_expect_mid_close(socket);
     AVS_UNIT_ASSERT_SUCCESS(avs_stream_cleanup((avs_stream_t **) &stream));
     avs_http_free(client);
 }
@@ -168,7 +168,7 @@ AVS_UNIT_TEST(http, init_https_auth2) {
     AVS_UNIT_ASSERT_EQUAL(stream->status, 0);
     AVS_UNIT_ASSERT_EQUAL(stream->redirect_count, 0);
     AVS_UNIT_ASSERT_NULL(stream->user_headers);
-    avs_unit_mocksock_expect_shutdown(socket);
+    avs_unit_mocksock_expect_mid_close(socket);
     AVS_UNIT_ASSERT_SUCCESS(avs_stream_cleanup((avs_stream_t **) &stream));
     avs_http_free(client);
 }